requirements.txt
fastapi==0.112.0
uvicorn==0.23.2
python-multipart==0.0.20
openai==1.3.0
python-dotenv==1.0.0
pillow==10.0.0
requests==2.31.0